Move and remove private error APIs
magic function is really only for the derive so move that to the
private module for now. The other function,
read_options_then_after_parse, was not documented and is also not
actually used by anything *in* binrw. If it was intended to be a
helper function it should not be in the error module anyway so it
is unlikely anyone ever found it there and decided to use it.
